implemented converting to json and exporting

+++ src/com/subtitles/sup/model/PGS.java
@@ -1,7 +1,7 @@
 package com.subtitles.sup.model;
 
 public class PGS {
-    public int presentationTimestamp;
+    public int presentationTimestamp; // timestamps have an accuracy of 90 kHz... devide by 90 to get the timestamp in milliseconds
     public int decodingTimestamp;
-    public short sizeOfSegment;
+    public int sizeOfSegment;
 }
